Add total count to publication

+++ GhostNetwork.Gateway.Facade/CommentsShort.cs
@@ -4,13 +4,13 @@ namespace GhostNetwork.Gateway.Facade
 {
     public class CommentsShort
     {
-        public CommentsShort(IEnumerable<PublicationComment> topComments, int totalCount)
+        public CommentsShort(IEnumerable<PublicationComment> topComments, long totalCount)
         {
             TopComments = topComments;
             TotalCount = totalCount;
         }
 
         public IEnumerable<PublicationComment> TopComments { get; }
-        public int TotalCount { get; }
+        public long TotalCount { get; }
     }
 }
